added funtionality in search, server and other changes

+++ apps/server/src/controllers/product.ts
@@ -1,10 +1,12 @@
 import { Request, Response } from "express";
+import mongoose from "mongoose";
 import {
   createProduct,
   updateProduct,
   deleteProduct,
   getAllProducts,
   getProductById,
+  getBySearchQuery,
 } from "../services/product";
 
 export const createProductController = async (req: Request, res: Response) => {
@@ -27,6 +29,10 @@ export const createProductController = async (req: Request, res: Response) => {
 export const updateProductController = async (req: Request, res: Response) => {
   const productId = req.params.id;
   const productData = req.body;
+  if (!mongoose.Types.ObjectId.isValid(productId)) {
+    return res.status(400).json({ error: "Invalid product id" });
+  }
+
   const updatedProduct = await updateProduct(productId, productData);
   if (!updatedProduct)
     return res.status(404).json({ error: "Product not found" });
@@ -38,6 +44,10 @@ export const updateProductController = async (req: Request, res: Response) => {
 
 export const deleteProductController = async (req: Request, res: Response) => {
   const productId = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(productId)) {
+    return res.status(400).json({ error: "Invalid product id" });
+  }
+
   const deletedProduct = await deleteProduct(productId);
   if (!deletedProduct)
     return res.status(404).json({ error: "Product not found" });
@@ -54,7 +64,20 @@ export const getAllProductsController = async (req: Request, res: Response) => {
 
 export const getProductByIdController = async (req: Request, res: Response) => {
   const productId = req.params.id;
+  if (!mongoose.Types.ObjectId.isValid(productId)) {
+    return res.status(400).json({ error: "Invalid product id" });
+  }
+
   const product = await getProductById(productId);
   if (!product) return res.status(404).json({ error: "Product not found" });
   res.json({ product });
+};
+
+export const getBySearchController = async (req: Request, res: Response) => {
+  const { name, category } = req.query as { name?: string; category?: string };
+  const price = req.query.price
+    ? parseInt(req.query.price as string)
+    : undefined;
+  const products = await getBySearchQuery({ name, category, price });
+  res.json({ products });
 };
